requirements.txt
beautifulsoup4
requests
lxml
sendgrid
python-dotenv